requirements.txt
-e .
aws-cdk.aws-lambda
aws-cdk.aws-lambda-event-sources
aws-cdk.aws-dynamodb
aws-cdk.aws-apigateway
aws-cdk.aws-sqs
boto3